Restructure pyISSM to include model module with io
Proposed restructure of pyISSM to include a model module, placing Model() class in 'core.py'. This allows more intuitive model io. Moved current model loading functions from data.io to model.io. data.io should be reserved for explicit handling of datasets external to ISSM models (e.g. data pool / functions to interact with external datasets).
